feat: add forum, login page, public controller, and update web modules
- Add ForumController, ForumPost, ForumComment models and repositories
- Add PublicController for public-facing endpoints
- Add forum.html and login.html static pages
- Update SecurityConfig, TransactionController, UserRepository
- Update static pages: about, admin, blog, dashboard, events, finances, index, profile
- Update api.js and style.css
- Update application.properties
- Update tukac-portal: Database, User model, LoginPanel, ManageEventsPanel, ManageFinancesPanel, pom.xml
- Add .gitignore to exclude build artifacts and DB files

+++ tukac-portal/src/main/java/com/tukac/ui/panels/ManageEventsPanel.java
@@ -199,7 +199,7 @@ public class ManageEventsPanel extends JPanel {
             pstmt.setString(4, timeField.getText().trim());
             pstmt.setString(5, locationField.getText().trim());
             pstmt.setInt(6, Integer.parseInt(capacityField.getText().trim().isEmpty() ? "0" : capacityField.getText().trim()));
-            pstmt.setInt(7, currentUser.getId());
+            pstmt.setLong(7, currentUser.getId());
             pstmt.executeUpdate();
             JOptionPane.showMessageDialog(this, "Event added!");
             clearForm();
